# feat[venom]: optimize `get_basic_block()`