--- a/vyper/venom/analysis/liveness.py
+++ b/vyper/venom/analysis/liveness.py
@@ -15,7 +15,7 @@ class LivenessAnalysis(IRAnalysis):
         self._reset_liveness()
         while True:
             changed = False
-            for bb in self.function.basic_blocks:
+            for bb in self.function.get_basic_blocks():
                 changed |= self._calculate_out_vars(bb)
                 changed |= self._calculate_liveness(bb)
 
@@ -23,7 +23,7 @@ class LivenessAnalysis(IRAnalysis):
                 break
 
     def _reset_liveness(self) -> None:
-        for bb in self.function.basic_blocks:
+        for bb in self.function.get_basic_blocks():
             bb.out_vars = OrderedSet()
             for inst in bb.instructions:
                 inst.liveness = OrderedSet()